services/tts-service/requirements.txt
Flask
Flask-Cors
piper-tts
gunicorn